Move RefreshVersionsMigrateTask from core to dependencies
