+++ plugins/dependencies/src/test/resources/migration.files/app/feature/build.gradle.kts
