+++ plugins/dependencies/src/test/resources/migration.files/buildSrc/src/main/kotlin/Deps.kt
