+++ plugins/dependencies/src/test/resources/migration.files/gradle/libraries.gradle
@@ -1,0 +1,1 @@
+package migration.files.gradle
