+++ plugins/dependencies/src/test/resources/migration.files/libraries.groovy
@@ -1,0 +1,1 @@
+package migration.files
