+++ plugins/dependencies/src/test/resources/migration.files/libs.gradle
@@ -1,0 +1,1 @@
+package migration.files
